Read 3D larval zebrafish template brain from Kunst et al. 2019

Read the template brain for a 6 day old larval zebrafish, generated by Kunst et al. 2019, straight from https://fishatlas.neuro.mpg.de. 3D triangular meshes are returned as a nat package hxsurf object, which mimics the Amira surface format. These can be be plotted in 3D using rgl and analysed with tools from the nat ecosystem. This incldue subseting by neuropil, i.e.. if you only want to visualise or analyse the forebrain.

fishatlas_read_brain(type = c("brain_regions", "outline", "fiber",
  "cell_bodies"), fishatlas_url = "https://fishatlas.neuro.mpg.de",
  surf = NULL)

Arguments

type
which version of the template brain you want to retreive. See details. Defaults to retreiving an expert segmnetation of the template brain into 51 neuropils

fishatlas_url
the web location of the fish brain atlas project

surf
Use if you want to add the downloaded brain to a hxsruf object already in your environment. Defaults to null, a new object is made from scratch

Value

a nat package hxsurf object, which mimics the Amira surface format, replete with metadata that can be accessed using $

Details

A single 3D brain object orits neuropils are read from .txt files hosted on https://fishatlas.neuro.mpg.de, and turned into a hxsurf object in R, which mimics the Amira surface format. As of May 2019, data available is:

outline an outline from the zebrafish brain

fiber a 3D mesh based on a tract segmentation of the brain

cell_bodies a 3D mesh based on segmenting cell bodies of the brain

brain_regions 51 expert annotated sub-regions of the brain. These include the major subdivisions beneth the forebrain, midbrain and hindbrain. Named in $Regionlist as 'major.neuropil_minor.neuropil_sub.region', e.g Forebrain_Diencephalon_Habenula

Note that since neurons are reconstructed from many different neural species, there is no 'standard' orientation between species, but within a species these neurons are registered to a template brain, usually using elastix.

References

Kunst, Michael, Eva Laurell, Nouwar Mokayes, Anna Kramer, Fumi Kubo, António M. Fernandes, Dominique Förster, Marco Dal Maschio, and Herwig Baier. 2019. “A Cellular-Resolution Atlas of the Larval Zebrafish Brain.” Neuron, May. https://doi.org/10.1016/j.neuron.2019.04.034.

See also

Examples

# NOT RUN {
## Lets have a look at these neuropils!
zfbrain = fishatlas_read_brain(type = "brain_regions")
plot3d(zfbrain)

## Maybe we just want to plot for forebrain
clear3d()
plot3d(subset(zfbrain, "Forebrain"), alpha = 0.5, col = "red")

## Maybe within the forebrain, we are interested in plotting the habenula
plot3d(subset(zfbrain, "Habenula"), alpha = 0.3, col = "red")

# }